Fix width blowout with wide content
In React Brand there are a few pages where the content on large screens
is wider than other pages. At certain breakpoints it also overlaps the
navigation.

This is a quirk of a flex child not collapsing where its children have
wider content than the flex calculation. The default min-width is auto,
and since there are children with width: 100%, auto corresponds
to max-content, which causes the blowout. To fix this we can set the
min-width to 0, which will make the children respect the flex calculation.

See https://css-tricks.com/flexbox-truncated-text/#aa-the-solution-is-min-width-0-on-the-flex-child